refactor(docs-site): extract web-style analytics wrapper for page views
Split the single gustoAnalytics client module into the layered wrapper
pattern used by web's standard-analytics: loading (loadAnalytics) is
separated from emission, page views go through a typed PageViewedEvent +
trackEvent entry point, and routeChangeListener is the reusable emit
primitive the host wires to Docusaurus' lifecycle and the consent bridge.

Loading stays the static-site CDN integration (WALKTHROUGH.md §3); wire
format is unchanged (GustoAnalytics.page).
